refactor: Rename states directory with view_states
