+++ lib/presentation/weather/view_models/weather_view_model.dart
@@ -1,7 +1,7 @@
 import 'package:flutter_training/application/weather/usecase_providers.dart';
 import 'package:flutter_training/domain/common/interfaces/result.dart';
 import 'package:flutter_training/domain/weather/entities/weather_target_entity.dart';
-import 'package:flutter_training/presentation/weather/states/weather_view_state.dart';
+import 'package:flutter_training/presentation/weather/view_states/weather_view_state.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 
 part 'weather_view_model.g.dart';
