+++ lib/presentation/weather/view_states/weather_view_state.freezed.dart
